fix(dcellar-web-ui): disable selection of SPs with invalid status in SPSelector
